chore: reorganise directory structure
