+++ src/components/Canvas.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Header from "./Header.tsx";
-import Footer from "./Footer.tsx";
+import Header from "./Header";
+import Footer from "./Footer";
 
 const Canvas = ({ children }) => (
   <>
