+++ src/components/PrivacyNote.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+
 import Modal from "@mui/material/Modal";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
