+++ src/pages/faq.tsx
