+++ src/styles/ResultView.css
